chore(module:breadcrumb): refactor breadcrumb
