+++ components/breadcrumb/public-api.ts
@@ -6,6 +6,6 @@
  * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
  */
 
-export * from './nz-breadcrumb-item.component';
-export * from './nz-breadcrumb.component';
-export * from './nz-breadcrumb.module';
+export * from './breadcrumb-item.component';
+export * from './breadcrumb.component';
+export * from './breadcrumb.module';
